requirements.txt
pydantic
python-dotenv
trafilatura
tenacity
whoosh
requests